MiniProject - CloudVault/requirements.txt
flask
python-dotenv
azure-storage-blob
supabase
google-generativeai
werkzeug
pypdf
python-docx
opencensus-ext-azure
